requirements_rag.txt
streamlit
groq
pandas
numpy
PyPDF2
beautifulsoup4
sentence-transformers
faiss-cpu 